requirements.txt
pandas>=2.0.0
clickhouse-connect>=0.6.0
python-dotenv>=1.0.0
